requirements.txt
Django==5.0
django-model-utils==4.3.1
djangorestframework==3.14.0
